Create universal_binary_counter.sv
Exercise for universal binary counter. It can be done first with full sequential and nested if in the ff. Or we can write an always_comb and have the nested if there. But I wanted to use the case (ctrl) inside. It is almost a hint the table of page 24 to use ?. But, apparently by chatGPT, casez also exist.